+++ env_example
@@ -1,0 +1,3 @@
+MONGO_URI=
+SYSTEM_EMAIL=
+SYSTEM_EMAIL_PASSWORD=
